pyproject.toml
[tool.poetry]
name = "ai-service"
version = "1.0.0"
description = "AI-powered code analysis service with Neo4j and Qdrant"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = "^0.34.0"
neo4j = "^5.28.0"
qdrant-client = "^1.14.0"
sentence-transformers = "^3.5.2"
tree-sitter = "^0.24.4"
gitpython = "^3.1.43"
pydantic = "^2.10.4"

[tool.poetry.dev-dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.25.2"
pytest-cov = "^6.0.0"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "--cov=src --cov-report=html --cov-report=term-missing"
